+++ tests/run.clj
@@ -61,6 +61,7 @@
 (require "tests/empty_list_test")
 (require "tests/error_path_test")
 (require "tests/bc_try_catch_test")
+(require "tests/bc_error_quality_test")
 (require "tests/bc_binding_test")
 (require "tests/bc_destructure_test")
 (require "tests/bc_closure_test")